+++ setup_bitswap.go
@@ -28,7 +28,11 @@ func setupBitswapExchange(ctx context.Context, cfg Config, h host.Host, cr routi
 
 	// Custom query manager with the content router and the host
 	// and our custom options to overwrite the default.
-	pqm, err := providerquerymanager.New(ctx, h, cr, providerquerymanager.WithMaxInProcessRequests(100))
+	pqm, err := providerquerymanager.New(ctx, h, cr,
+		providerquerymanager.WithMaxInProcessRequests(cfg.RoutingMaxRequests),
+		providerquerymanager.WithMaxProviders(cfg.RoutingMaxProviders),
+		providerquerymanager.WithMaxTimeout(cfg.RoutingMaxTimeout),
+	)
 	if err != nil {
 		panic(err)
 	}